+++ CHANGELOG.md
@@ -8,10 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 New release of the HyperDbg Debugger.
 
 ### Added
+- The **!monitor** command now supports 'execution' interception ([link](https://docs.hyperdbg.org/commands/extension-commands/monitor))
 - **!crwrite** - Control Register Modification Event ([link](https://docs.hyperdbg.org/commands/extension-commands/crwrite))
-
-## [0.3.1.0] - 2023-06-29
-New release of the HyperDbg Debugger.
 
 ### Changed
 - The '.start' command's mechanism for finding the entrypoint is changed to address issues ([link](https://github.com/HyperDbg/HyperDbg/commit/c64cb39ebe066865a85e9dd4b696295260a81f18))